scripts/requirements.txt
facebook-sdk>=3.1.0
python-dateutil>=2.8.2 